feat: Implement sell report screen and enhance item management
This commit introduces a new "Sell Reports" screen that allows users to view and filter past bills by date or bill ID. It also includes several improvements to item management:

- **Sell Reports Screen:**
    - Displays a list of bills, grouped by date in descending order.
    - Provides a search bar to filter bills by bill ID or date (yyyy-mm-dd).
    - Tapping on a bill navigates to the bill details screen.
    - Shows an icon indicating if a bill is paid or pending.
- **Item Quantity:** Item quantity is now handled as a `double` to support fractional values.
- **Item Provider Enhancements:**
    - `SelectedItemsNotifier` now includes methods to `addItem`, `deleteItem`, and `editItem`.
- **UI Improvements:**
    - Drawer navigation updated to include "Sell list" and "Report" which navigate to the new Sell Reports screen.
    - Item card in `SelectItem` screen now displays item price, quantity, and total.
    - Long-pressing an item in the `SelectItem` screen now opens the item edit bottom sheet.
    - Item image display in `SelectItem` and `AddItem` has been improved.
    - `AddItem` bottom sheet now supports adding multiple images for an item.
    - `AddItem` bottom sheet UI has been refined with better spacing and input field styling.
- **Bug Fixes:**
    - Corrected status display for paid bills in `HomeScreen`.
    - Ensured selected items are correctly passed to `BillDetailsScreen`.
    - PDF generator now correctly calculates total quantity as a number.

+++ lib/model/item_model.dart
@@ -1,7 +1,7 @@
 class Items {
   final String id;
   final String productName;
-  final int quantity;
+  final double quantity;
   final double price;
   final String? unit;
   final String category;
@@ -50,7 +50,7 @@ class Items {
   Items copyWith({
     String? id,
     String? productName,
-    int? quantity,
+    double? quantity,
     double? price,
     String? image,
     String? unit,
